:sparkles: Add reduce.js to split reducer from store.js
